feat: added getPreviousWords functionality
These changes allow for the querying of previous words.

Since I am currently not working with an external API (to keep server cost to a minimum), instead I will keep a record of the dates that I have added words for.

That way I only perform the fetches I need to, rather than running X amount of fetches to only get a small number of data.