feat: added startGameWithPreviousWord function
This adds a new method which can be used to control the game state, essentially resetting the game with an arbitrary word; in this case, a word selected from the card.

The condition checks in the main useEffect in GameContextProvider means we don't have to make any changes into how we load the game there; instead we just ensure the dependency array is empty so that the core attempt to load the daily game only happens once.